refactor: separate choice answer form
